hotfix: pinia-plugin-persistedstate 추가 & 헤더에 유저 정보 반응형으로 가져오기
